Added functions Alternative_Connected_Components
In get_tail_split the code that was used to find the connected components has been replaced with two new functions called Alternative_Connected_Components (in various_functions.cpp). These functions do not change the algorithm, they are just for readability. There are two functions, with the same name, that take different parameters and return different values; the first is the basic one used for the remain vector, the second one takes as input the current connected components found in the previous loop and updates those by overwriting the vector.